Create views.py
In a Django web application, the views.py file defines the functions or classes that handle HTTP requests from the client and return an HTTP response. 
The views are the logic behind the application's user interface, determining what data is displayed to the user and how it is presented.

In other words, the views.py file is an essential part of a Django application as it acts as the bridge between the front-end (HTML templates) and back-end (database, models) of the application. It receives requests from the user, processes them using the necessary logic, retrieves or modifies data from the database through models, and returns an appropriate response.